requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
streamlit>=1.35.0
pydeck>=0.9.0
pandas>=2.0.0
folium>=0.17.0
streamlit-folium>=0.22.0
